Added fix so it will work on R2009b.
git-svn-id: https://matano1-tp.haifa.ibm.com/svn/taco@273 4c94cbd0-a001-4ae1-9522-79914504ef8a

+++ Feature Extraction/InstancesSet.m
@@ -126,7 +126,7 @@ methods (Static)
             R = map(key);
         else
             % a new feature
-            newMappedNumber = map.Count + 1;
+            newMappedNumber = uint32(map.Count) + 1;
             map(key) = newMappedNumber; %#ok<NASGU>
             R = newMappedNumber ;
         end
